mental_health_chatbot/requirements.txt
flask
google-genai
python-dotenv


